Replace int with size_t for size/count/stride/index variables across csrc/ (#6)
* Replace int with size_t for size/count/stride/index variables in cache.cpp

Change local variables assigned from .size() calls, kernel class constructor
parameters, kernel class member variables, and loop variables to use size_t
instead of int for variables representing sizes, dimensions, counts, strides,
indices, block sizes, and head sizes. This avoids signed/unsigned comparison
bugs.


* Replace int with size_t for size/dimension/count variables in layernorm kernels

Change variables representing sizes, dimensions, counts, and loop indices
from int to size_t in csrc/layernorm.cpp and csrc/layernorm_quant.cpp.

This affects:
- Kernel class constructor parameters and member variables
- Local variables (hidden_size, num_tokens, num_dims, etc.)
- Loop variables iterating over sizes/dimensions
- Fix std::min calls with static_cast<size_t>(1024)
- Fix std::gcd calls with static_cast<size_t>(...)

Preserved: template non-type params (int NUM_DIMS, int VEC_SIZE),
int64_t variables, and inner j-loops bounded by VEC_SIZE.



* Replace int with size_t for size/dimension/count variables in pos_encoding and fused_qknorm_rope kernels

- pos_encoding_kernels.cpp: Changed positions_ndim, query_hidden_size,
  key_hidden_size, num_heads, num_kv_heads, rot_dim, seq_dim_idx,
  query_ndim, and kernel class member vars (head_size, num_heads,
  num_kv_heads, rot_dim) from int to size_t. Updated loop/index vars
  (embed_dim, nq, nk, head_idx, rot_offset, x_index, y_index) and
  fixed std::min<int64_t> to std::min<size_t>.

- fused_qknorm_rope.cpp: Changed kernel class constructor params and
  member vars (num_heads_q, num_heads_k, num_heads_v, num_tokens,
  rotary_dim) from int to size_t. Updated all loop vars, index vars
  (total_qk_heads, token_idx, local_head_idx, head_idx, num_heads,
  offset_warp, offset_thread, embed_dim, rotary_lanes, pair_offset,
  dim_idx, half_dim), and launch function params. Changed static_cast<int>
  to static_cast<size_t> at call site.

Preserved: int token_idx in rotary_embedding (used with int64_t strides),
template non-type params (int head_dim), int64_t variables, and
constexpr int constants (block_size, sgs_per_wg, SUB_GROUP_SIZE,
NUM_ELEMS_PER_THREAD).



* Replace int with size_t for size/dimension/count variables in moe kernels (partial)

Changed int → size_t for variables representing sizes, dimensions, counts,
strides, and indices in 4 files:

- fused_moe_prologue.cpp: experts_per_token
- init_expert_map.cpp: num_experts, ep_rank, ep_size (params, members, launcher)
- moe_gather.cpp: num_experts, num_tokens, hidden_size, topk, loop vars
- remap_hidden_states.cpp: RowsPerExpertCount get_nd_range params (partial)

Note: This is a partial change. The following files still need updates:
- topk.cpp
- moe_align_sum_kernels.cpp
- remap_hidden_states.cpp (remaining changes)
- fused_grouped_topk.cpp
- grouped_topk.cpp
- fused_moe_prologue.hpp



* Replace int with size_t for size/count/stride/index variables across csrc/ to fix signed/unsigned comparison risks

Agent-Logs-Url: https://github.com/jikunshang/vllm-xpu-kernels/sessions/77ac7b1b-2a91-4f87-9031-10fbf67b346f


---------

+++ csrc/moe/remap_hidden_states.cpp
@@ -13,8 +13,8 @@ class RowsPerExpertCount {
       int64_t* expert_first_token_offset,
       int64_t* topk_ids,
       int* unpermuted_row_to_permuted_row,
-      const int num_rows,
-      const int TopK)
+      const size_t num_rows,
+      const size_t TopK)
       : expert_map(expert_map),
         expert_first_token_offset(expert_first_token_offset),
         topk_ids(topk_ids),
